GCP/function-2/requirements.txt
functions-framework==3.*
boto3
google-cloud-storage